pyproject.toml
[project]
name = "fastapi-fs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "bcrypt==4.0.1",
    "black>=25.9.0",
    "fastapi[standard]>=0.118.0",
    "greenlet>=3.2.4",
    "passlib>=1.7.4",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "redis>=7.0.1",
    "sqlmodel>=0.0.27",
]
